rentableapi/requirements.txt
databases[asyncpg]==0.9.0
dependency-injector==4.48.3
fastapi==0.115.4
numpy==2.1.3
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
SQLAlchemy==2.0.36
uvicorn==0.32.0
greenlet==3.1.1
bcrypt==4.2.1